Refactored board related fnctions
